+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # RDS-NG - Changelog
 
+## v1.0.4 - 2025-07-25
+
+### Improvements
+
+- Show a hint if no connections have been added in the projects list
+
+### Fixes
+
+- The wrong root directory is used for data annotation
+
 ## v1.0.3 - 2025-07-22
 
 ### Improvements